+++ yandex/resource_yandex_cdn_resource_test.go
@@ -530,6 +530,35 @@ func TestAccCDNResource_optionStaticRequestHeaders(t *testing.T) {
 	})
 }
 
+func TestAccCDNResource_optionSecureKey(t *testing.T) {
+	folderID := getExampleFolderID()
+
+	groupName := fmt.Sprintf("tf-test-cdn-resource-%s", acctest.RandString(10))
+	resourceCName := fmt.Sprintf("cdn-tf-test-%s.yandex.net", acctest.RandString(4))
+
+	var cdnResource cdn.Resource
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckCDNResourceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccCDNResource_optionSecureKey(groupName, resourceCName),
+				Check: resource.ComposeTestCheckFunc(
+					testCDNResourceExists("yandex_cdn_resource.foobar_resource", &cdnResource),
+					resource.TestCheckResourceAttr("yandex_cdn_resource.foobar_resource", "cname", resourceCName),
+					resource.TestCheckResourceAttr("yandex_cdn_resource.foobar_resource", "folder_id", folderID),
+					resource.TestCheckResourceAttr("yandex_cdn_resource.foobar_resource", "options.#", "1"),
+					resource.TestCheckResourceAttr("yandex_cdn_resource.foobar_resource", "options.0.secure_key", "testsecurekey"),
+					resource.TestCheckResourceAttr("yandex_cdn_resource.foobar_resource", "options.0.enable_ip_url_signing", "true"),
+					testAccCheckCreatedAtAttr("yandex_cdn_resource.foobar_resource"),
+				),
+			},
+		},
+	})
+}
+
 func testSweepCDNResource(_ string) error {
 	conf, err := configForSweepers()
 	if err != nil {
@@ -703,6 +732,21 @@ resource "yandex_cdn_resource" "foobar_resource" {
 
 	options {
 		ignore_cookie = true
+	}
+}
+`, resourceCNAME)
+}
+
+func testAccCDNResource_optionSecureKey(groupName, resourceCNAME string) string {
+	return makeGroupResource(groupName) + fmt.Sprintf(`
+resource "yandex_cdn_resource" "foobar_resource" {
+	cname = "%s"
+
+	origin_group_id = "${yandex_cdn_origin_group.foo_cdn_group.id}"
+
+	options {
+		secure_key = "testsecurekey"
+		enable_ip_url_signing = true
 	}
 }
 `, resourceCNAME)